+++ app/dashboard/page.tsx
@@ -2,6 +2,7 @@
 
 import FavCountGraph from "@/components/cards/fav-count.linegraph.card";
 import TrendingHashtagCard from "@/components/cards/trending-hastag.card";
+import TweetAreaGraphCard from "@/components/cards/tweet-areagraph.card";
 import TweetGraphCard from "@/components/cards/tweet-linegraph.card";
 import { useAppSelector } from "@/store/store";
 
@@ -10,10 +11,10 @@ export default function Dashboard() {
 
   return (
     <>
-      {/* <TweetGraphCard username={username} reply={false} limit={1} /> */}
+      <TweetAreaGraphCard username={username} reply={false} limit={1} />
       {/* <TrendingHashtagCard woeid={1}/> */}
       {/* <CountGraphCard username={username} reply={false} limit={1}/> */}
-      <FavCountGraph username={username} reply={false} limit={1}/>
+      {/* <FavCountGraph username={username} reply={false} limit={1}/> */}
 
     </>
   );
